backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
packaging==25.0
psycopg==3.2.12
psycopg-binary==3.2.12
python-dotenv==1.2.1
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.3
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.11.0
